use Wpcom_Products::get_site_current_purchases to reduce requests
