+++ projects/packages/my-jetpack/src/products/class-starter.php
@@ -155,6 +155,26 @@ class Starter extends Module_Product {
 	}
 
 	/**
+	 * Checks whether the current plan (or purchases) of the site already supports the product
+	 *
+	 * @return boolean
+	 */
+	public static function has_required_plan() {
+		$purchases_data = Wpcom_Products::get_site_current_purchases();
+		if ( is_wp_error( $purchases_data ) ) {
+			return false;
+		}
+		if ( is_array( $purchases_data ) && ! empty( $purchases_data ) ) {
+			foreach ( $purchases_data as $purchase ) {
+				if ( 0 === strpos( $purchase->product_slug, 'jetpack_starter' ) ) {
+					return true;
+				}
+			}
+		}
+		return false;
+	}
+
+	/**
 	 * Checks whether product is a bundle.
 	 *
 	 * @return boolean True
@@ -170,15 +190,6 @@ class Starter extends Module_Product {
 	 */
 	public static function get_supported_products() {
 		return array( 'backup', 'anti-spam' );
-	}
-
-	/**
-	 * Return all the products that support the starter product.
-	 *
-	 * @return Array Product slugs
-	 */
-	public static function get_products() {
-		return array( 'jetpack_starter' );
 	}
 
 	/**
